requirements.txt
pandas
fmpsdk
duckdb
matplotlib
pyscript-ltk
